backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=1.4.47
psycopg2-binary>=2.9.6
openai>=0.27.8
python-dotenv>=1.0.0
langchain>=0.0.217
pandas>=1.5.3
